All classes were divided into separate files. A new tactic for moving 4 trains was added. Now the game moves automatically using thread
