+++ py_modules/colors.py
@@ -1,0 +1,12 @@
+BLACK = (0, 0, 0)
+WHITE = (255, 255, 255)
+YELLOW = (225, 225, 0)
+BLUE = (0, 0, 255)
+DARK_BLUE = (0, 0, 100)
+RED = (200, 0, 0)
+DARK_RED = (100, 0, 0)
+GREEN = (0, 200, 0)
+DARK_GREEN = (0, 100, 0)
+colors = [YELLOW, RED, GREEN, BLUE, WHITE]
+type_colors = {"armor": [BLUE, DARK_BLUE], "population": [RED, DARK_RED], "product": [GREEN, DARK_GREEN]}
+train_colors = {1: (255, 255, 120), 2: (120, 255, 255), 3: (255, 120, 255), 4: (128, 128, 128)}
